// src/webs/CSSPWebToolsAng/client-apps/src/app/models/generated/Contact.model.ts
/*
 * Auto generated from C:\CSSPTools\src\codegen\Tests\_package\netcoreapp3.1\testhost.exe
 *
 * Do not edit this file.
 *
 */

import { LastUpdate } from './LastUpdate.model';
import { ContactTitleEnum } from '../../enums/generated/ContactTitleEnum';

export class Contact extends LastUpdate {
    ContactID: number;
    ContactTitle?: ContactTitleEnum;
    ContactTVItemID: number;
    Disabled: boolean;
    EmailValidated: boolean;
    FirstName: string;
    Id: string;
    Initial: string;
    IsAdmin: boolean;
    IsNew: boolean;
    LastName: string;
    LastUpdateContactTVItemID: number;
    LastUpdateDate_UTC: Date;
    LoginEmail: string;
    SamplingPlanner_ProvincesTVItemID: string;
    Token: string;
    WebName: string;
}