// src/webs/CSSPWebToolsAng/client-apps/src/app/models/generated/SearchResult.model.ts
/*
 * Auto generated from C:\CSSPTools\src\codegen\_package\netcoreapp3.1\AngularModelsGenerated.exe
 *
 * Do not edit this file.
 *
 */

import { TVItem } from './TVItem.model';
import { TVItemLanguage } from './TVItemLanguage.model';

export class SearchResult {
    TVItem: TVItem;
    TVItemLanguage: TVItemLanguage;
}