// src/webs/CSSPWebToolsAng/client-apps/src/app/models/generated/WebArea.model.ts
/*
 * Auto generated from C:\CSSPTools\src\codegen\_package\netcoreapp3.1\AngularModelsGenerated.exe
 *
 * Do not edit this file.
 *
 */

import { LastUpdate } from './LastUpdate.model';
import { MapInfo } from './MapInfo.model';
import { MapInfoPoint } from './MapInfoPoint.model';
import { TVFileLanguage } from './TVFileLanguage.model';
import { TVFile } from './TVFile.model';
import { TVItem } from './TVItem.model';
import { TVItemLanguage } from './TVItemLanguage.model';
import { TVItemLink } from './TVItemLink.model';
import { TVItemStat } from './TVItemStat.model';

export class WebArea extends LastUpdate {
    MapInfoList: MapInfo[];
    MapInfoPointList: MapInfoPoint[];
    MapInfoPointSectorList: MapInfoPoint[];
    MapInfoSectorList: MapInfo[];
    TVFileLanguageList: TVFileLanguage[];
    TVFileList: TVFile[];
    TVItem: TVItem;
    TVItemLanguageList: TVItemLanguage[];
    TVItemLanguageSectorList: TVItemLanguage[];
    TVItemLinkList: TVItemLink[];
    TVItemSectorList: TVItem[];
    TVItemStatList: TVItemStat[];
    TVItemStatSectorList: TVItemStat[];
}