// src/webs/CSSPWebToolsAng/client-apps/src/app/models/generated/WebContact.model.ts
/*
 * Auto generated from C:\CSSPTools\src\codegen\_package\netcoreapp3.1\AngularModelsGenerated.exe
 *
 * Do not edit this file.
 *
 */

import { LastUpdate } from './LastUpdate.model';
import { Contact } from './Contact.model';

export class WebContact extends LastUpdate {
    ContactList: Contact[];
}