// src/webs/CSSPWebToolsAng/client-apps/src/app/models/generated/WebCountry.model.ts
/*
 * Auto generated from C:\CSSPTools\src\codegen\_package\netcoreapp3.1\AngularModelsGenerated.exe
 *
 * Do not edit this file.
 *
 */

import { EmailDistributionListContactLanguage } from './EmailDistributionListContactLanguage.model';
import { EmailDistributionListContact } from './EmailDistributionListContact.model';
import { EmailDistributionListLanguage } from './EmailDistributionListLanguage.model';
import { EmailDistributionList } from './EmailDistributionList.model';
import { TVItemModel } from './TVItemModel.model';
import { WebBase } from './WebBase.model';

export class WebCountry {
    EmailDistributionListContactLanguageList: EmailDistributionListContactLanguage[];
    EmailDistributionListContactList: EmailDistributionListContact[];
    EmailDistributionListLanguageList: EmailDistributionListLanguage[];
    EmailDistributionListList: EmailDistributionList[];
    TVItemModel: TVItemModel;
    TVItemParentList: WebBase[];
    TVItemProvinceList: WebBase[];
}