// src/webs/CSSPWebToolsAng/client-apps/src/app/models/generated/WebProvince.model.ts
/*
 * Auto generated from C:\CSSPTools\src\codegen\_package\netcoreapp3.1\AngularModelsGenerated.exe
 *
 * Do not edit this file.
 *
 */

import { SamplingPlan } from './SamplingPlan.model';
import { WebBase } from './WebBase.model';
import { TVItemModel } from './TVItemModel.model';

export class WebProvince {
    SamplingPlanList: SamplingPlan[];
    TVItemAreaList: WebBase[];
    TVItemModel: TVItemModel;
    TVItemParentList: WebBase[];
}